refactor: rename ContextEventHandler to ContextPublisher
