+++ sdk/absmartly.py
@@ -10,7 +10,7 @@ from sdk.context_config import ContextConfig
 from sdk.context_event_logger import ContextEventLogger
 from sdk.default_audience_deserializer import DefaultAudienceDeserializer
 from sdk.default_context_data_provider import DefaultContextDataProvider
-from sdk.default_context_event_handler import DefaultContextEventHandler
+from sdk.default_context_publisher import DefaultContextPublisher
 from sdk.default_http_client import DefaultHTTPClient
 from sdk.default_http_client_config import DefaultHTTPClientConfig
 from sdk.default_variable_parser import DefaultVariableParser
@@ -81,7 +81,7 @@ class ABsmartly:
 
             if self.context_event_handler is None:
                 self.context_event_handler = \
-                    DefaultContextEventHandler(self.client)
+                    DefaultContextPublisher(self.client)
 
         if self.variable_parser is None:
             self.variable_parser = DefaultVariableParser()
