+++ sdk/context_publisher.py
@@ -1,0 +1,14 @@
+from abc import abstractmethod
+from concurrent.futures import Future
+from typing import Optional
+
+from sdk.json.context_data import ContextData
+from sdk.json.publish_event import PublishEvent
+
+
+class ContextPublisher:
+
+    @abstractmethod
+    def publish(self, context, event: PublishEvent) -> \
+            Future[Optional[ContextData]]:
+        raise NotImplementedError
