+++ sdk/default_context_event_handler.py
@@ -1,19 +1,16 @@
-from concurrent.futures import Future
-from typing import Optional
+import warnings
 
 from sdk.client import Client
-from sdk.context import Context
-from sdk.context_event_handler import ContextEventHandler
-from sdk.json.context_data import ContextData
-from sdk.json.publish_event import PublishEvent
+from sdk.default_context_publisher import DefaultContextPublisher
 
 
-class DefaultContextEventHandler(ContextEventHandler):
+class DefaultContextEventHandler(DefaultContextPublisher):
+    """Deprecated: Use DefaultContextPublisher instead."""
 
     def __init__(self, client: Client):
-        self.client = client
-
-    def publish(self,
-                context: Context,
-                event: PublishEvent) -> Future[Optional[ContextData]]:
-        return self.client.publish(event)
+        warnings.warn(
+            "DefaultContextEventHandler is deprecated, use DefaultContextPublisher instead.",
+            DeprecationWarning,
+            stacklevel=2
+        )
+        super().__init__(client)
